Move assert_regex_format to validity module and update documentation
