+++ README.md
@@ -36,6 +36,8 @@ from data_quality_quick.validate_formats import check_type_format
   - **assert_that_dataframe_is_empty**: Check if a DataFrame is empty.
 - **Validity**
   - **assert_that_there_are_not_nulls**: Checks for null values in a specified column of a DataFrame.
+  - **assert_regex_format**: Validates whether all values in the specified column of a DataFrame match the format
+    defined by the provided regular expression, including null and empty values.
 - **Consistency**
   - **assert_that_there_are_not_duplicates**: Checks for duplicate values in the specified primary key column of a DataFrame.
   - **assert_that_columns_values_match** :  Check if all values in column2 of df2 are present in column1 of df1.
@@ -43,6 +45,5 @@ from data_quality_quick.validate_formats import check_type_format
   - **assert_that_type_value**: Check if all non-null entries in a specified column of a DataFrame are of the specified data type.
   - **assert_that_values_in_catalog**:  Checks whether all values in the specified column of a DataFrame are present
     in a catalog (list of values).
-  - **assert_regex_format**: Validates whether all values in the specified column of a DataFrame match the format
-    defined by the provided regular expression, including null and empty values.
+
